Fix comparison to resolve CI errors.

+++ Modules/ModelFitUI/Qmitk/QmitkFitParameterModel.cpp
@@ -272,7 +272,7 @@ data(const QModelIndex& index, int role) const
               else
               {
                 QString concatenatedValues;
-                for (int i = 0; i < currentFit->staticParamMap.Get(staticParamName).size(); i++)
+                for (long unsigned int i = 0; i < currentFit->staticParamMap.Get(staticParamName).size(); i++)
                 {
                   auto value = currentFit->staticParamMap.Get(staticParamName).at(i);
                   concatenatedValues += QString::number(value) + ", ";
